%%
%%
%%

\name{pistar-package}
\alias{pistar-package}

\docType{package}

\title{
	Rudas, Clogg, and Lindsay Mixture Index of Fit
}
\description{
	Functions for the Rudas, Clogg and Lindsay \eqn{\pi^{*}}{pi*} mixture index of fit.
}
\details{
	\tabular{ll}{
		Package: 	\tab pistar		\cr
		Type: 		\tab Package	\cr
		Version: 	\tab 0.5.2.3	\cr
		Date: 		\tab 2014-12-12	\cr
		License: 	\tab GPL (>= 2)	\cr
	}
	Functions for the Rudas, Clogg and Lindsay \eqn{\pi^{*}}{pi*} mixture index of fit.
}
\author{
	Juraj Medzihorsky \email{juraj.medzihorsky@gmail.com}
}
\note{
	The author would like to thank Tamas Rudas for invaluable advice in the development of this package.

	Work on the package was supported in part by Central European University grant BPF/5182/20123/R.
}
\references{}

\keyword{package}
\keyword{pistar}
\seealso{}
\examples{}
